Consolodated function files, refactored parser and ops, removed superfluous dlist functions
