--- 5-functions.c
@@ -1,15 +1,0 @@
-#include "monty.h"
-
-/**
-* nop - No operation, do nothing
-* @stack: Head of stack
-* @linenumber: Line number of monty file
-*/
-
-void nop(stack_t **stack, unsigned int linenumber)
-{
-	(void)stack;
-	(void)linenumber;
-
-	;
-}
